compliance with sonar lint

+++ src/main/java/com/james/io/search/IntegerSquareRootSolution.java
@@ -5,8 +5,8 @@ public class IntegerSquareRootSolution {
     public static int searchIntegerSquareRoot(int k) {
         int i = 0;
 
-        while (Math.pow(i, 2) < k) {
-            if (Math.pow(i + 1, 2) > k) {
+        while (Math.pow((double) i, 2) < k) {
+            if (Math.pow((double) i + 1, 2) > k) {
                 break;
             }
             i++;
